src/main/java/edu/baker/finalproject/BigDecimalUtils.java
package edu.baker.finalproject;

import java.math.BigDecimal;
import java.math.MathContext;
import java.math.RoundingMode;

/**
 * Utility class that provides BigDecimal-based mathematical operations.
 */
public class BigDecimalUtils {

    /**
     * Adds two BigDecimal numbers.
     * @param a First number.
     * @param b Second number.
     * @param mc MathContext for precision.
     * @return Sum of a and b.
     */
    public static BigDecimal add(BigDecimal a, BigDecimal b, MathContext mc) {
        // return a.add(b, mc);
        return BigDecimal.ZERO;
    }

    /**
     * Subtracts the second BigDecimal number from the first.
     * @param a First number.
     * @param b Second number.
     * @param mc MathContext for precision.
     * @return a - b.
     */
    public static BigDecimal subtract(BigDecimal a, BigDecimal b, MathContext mc) {
        // return a.subtract(b, mc);
        return BigDecimal.ZERO;
    }

    /**
     * Multiplies two BigDecimal numbers.
     * @param a First number.
     * @param b Second number.
     * @param mc MathContext for precision.
     * @return a * b.
     */
    public static BigDecimal multiply(BigDecimal a, BigDecimal b, MathContext mc) {
        // return a.multiply(b, mc);
        return BigDecimal.ZERO;
    }

    /**
     * Divides the first BigDecimal number by the second.
     * @param a First number.
     * @param b Second number.
     * @param mc MathContext for precision.
     * @return a / b.
     */
    public static BigDecimal divide(BigDecimal a, BigDecimal b, MathContext mc) {
        // return a.divide(b, mc);
        return BigDecimal.ZERO;
    }

    /**
     * Returns the minimum of two BigDecimal numbers.
     * @param a First number.
     * @param b Second number.
     * @return Minimum of a and b.
     */
    public static BigDecimal min(BigDecimal a, BigDecimal b) {
        // return a.min(b);
        return BigDecimal.ZERO;
    }

    /**
     * Returns the maximum of two BigDecimal numbers.
     * @param a First number.
     * @param b Second number.
     * @return Maximum of a and b.
     */
    public static BigDecimal max(BigDecimal a, BigDecimal b) {
        // return a.max(b);
        return BigDecimal.ZERO;
    }

    /**
     * Returns the absolute value of a BigDecimal number.
     * @param a The number.
     * @return Absolute value of a.
     */
    public static BigDecimal abs(BigDecimal a) {
        // return a.abs();
        return BigDecimal.ZERO;
    }

    /**
     * Raises a BigDecimal to the power of a double exponent.
     * @param base The base number.
     * @param exponent The exponent.
     * @param mc MathContext for precision.
     * @return base raised to the power of exponent.
     */
    public static BigDecimal pow(BigDecimal base, BigDecimal exponent, MathContext mc) {
        // return new BigDecimal(Math.pow(base.doubleValue(), exponent), mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the square root of a BigDecimal number.
     * @param value The number to find the square root of.
     * @param mc MathContext for precision.
     * @return Square root of value.
     */
    public static BigDecimal sqrt(BigDecimal value, MathContext mc) {
        // return value.sqrt(mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the sine of a BigDecimal angle in radians.
     * @param value The angle in radians.
     * @param mc MathContext for precision.
     * @return Sine of value.
     */
    public static BigDecimal sin(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.sin(value.doubleValue()), mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the cosine of a BigDecimal angle in radians.
     * @param value The angle in radians.
     * @param mc MathContext for precision.
     * @return Cosine of value.
     */
    public static BigDecimal cos(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.cos(value.doubleValue()), mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the tangent of a BigDecimal angle in radians.
     * @param value The angle in radians.
     * @param mc MathContext for precision.
     * @return Tangent of value.
     */
    public static BigDecimal tan(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.tan(value.doubleValue()), mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the arcsine of a BigDecimal value.
     * @param value The value.
     * @param mc MathContext for precision.
     * @return Arcsine of value.
     */
    public static BigDecimal asin(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.asin(value.doubleValue()), mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the arccosine of a BigDecimal value.
     * @param value The value.
     * @param mc MathContext for precision.
     * @return Arccosine of value.
     */
    public static BigDecimal acos(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.acos(value.doubleValue()), mc);
        return BigDecimal.ZERO;
    }

    /**
     * Computes the arctangent of a BigDecimal value.
     * @param value The value.
     * @param mc MathContext for precision.
     * @return Arctangent of value.
     */
    public static BigDecimal atan(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.atan(value.doubleValue()), mc);
        return BigDecimal.ZERO;

    }

    /**
     * Computes the natural logarithm of a BigDecimal value.
     * @param value The value.
     * @param mc MathContext for precision.
     * @return Natural logarithm of value.
     */
    public static BigDecimal log(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.log(value.doubleValue()), mc);
        return BigDecimal.ZERO;

    }

    /**
     * Computes the base-10 logarithm of a BigDecimal value.
     * @param value The value.
     * @param mc MathContext for precision.
     * @return Base-10 logarithm of value.
     */
    public static BigDecimal log10(BigDecimal value, MathContext mc) {
        // return new BigDecimal(Math.log10(value.doubleValue()), mc);
        return BigDecimal.ZERO;

    }

    /**
     * Rounds down a BigDecimal value to the nearest integer.
     * @param value The value to round down.
     * @return Rounded down value.
     */
    public static BigDecimal floor(BigDecimal value) {
        // Set the scale to 0 and use RoundingMode.FLOOR for rounding down.
        return value.setScale(0, RoundingMode.FLOOR);
    }

    /**
     * Rounds up a BigDecimal value to the nearest integer.
     * @param x
     * @return Rounded up value.
     */
    public static BigDecimal ceil(BigDecimal x) {
        // return value.setScale(0, RoundingMode.CEILING);
        double doubleValue = x.doubleValue();
        double ceilValue = Math.ceil(doubleValue);
        return new BigDecimal(ceilValue);
    }

    /**
     * Rounds a BigDecimal value to the nearest integer.
     * @param value The value to round.
     * @return Rounded value.
     */
    public static BigDecimal round(BigDecimal value) {
        // return value.setScale(0, RoundingMode.HALF_UP);
        return BigDecimal.ZERO;
    }
    
    /**
    * Returns the reciprocal of a BigDecimal number.
    * @param x The number.
    * @param mc MathContext for precision.
    * @return Reciprocal of x.
    */
   public static BigDecimal recip(BigDecimal x, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Truncates a BigDecimal number.
    * @param x The number.
    * @param mc MathContext for precision.
    * @return Truncated value of x.
    */
   public static BigDecimal trunc(BigDecimal x, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Raises a BigDecimal to the power of another BigDecimal exponent.
    * @param base The base number.
    * @param power The exponent.
    * @param mc MathContext for precision.
    * @return base raised to the power of exponent.
    */
   public static BigDecimal ipow(BigDecimal base, BigDecimal power, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Raises a BigDecimal to the power of a long exponent.
    * @param base The base number.
    * @param ipower The exponent.
    * @param mc MathContext for precision.
    * @return base raised to the power of exponent.
    */
   public static BigDecimal ipow(BigDecimal base, long ipower, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Computes the root of a BigDecimal number using another BigDecimal as the root value.
    * @param base The base number.
    * @param root The root value.
    * @param mc MathContext for precision.
    * @return base to the power of 1/root.
    */
   public static BigDecimal iroot(BigDecimal base, BigDecimal root, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Computes the root of a BigDecimal number using a long as the root value.
    * @param base The base number.
    * @param iroot The root value.
    * @param mc MathContext for precision.
    * @return base to the power of 1/root.
    */
   public static BigDecimal iroot(BigDecimal base, long iroot, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Computes 10 raised to the power of a BigDecimal number.
    * @param x The exponent.
    * @param mc MathContext for precision.
    * @return 10 raised to the power of x.
    */
   public static BigDecimal exp10(BigDecimal x, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Computes the arctangent of the quotient of its arguments.
    * @param y The y-coordinate.
    * @param x The x-coordinate.
    * @param mc MathContext for precision.
    * @return The angle whose tangent is y/x.
    */
   public static BigDecimal atan2(BigDecimal y, BigDecimal x, MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Returns the mathematical constant e.
    * @param mc MathContext for precision.
    * @return The value of e.
    */
   public static BigDecimal e(MathContext mc) {
        return new BigDecimal(Math.exp(1), mc);
   }

   /**
    * Returns the mathematical constant pi.
    * @param mc MathContext for precision.
    * @return The value of pi.
    */
   public static BigDecimal pi(MathContext mc) {
       return BigDecimal.ZERO;
   }

   /**
    * Computes the factorial of a BigDecimal number.
    * @param x The number.
    * @param mc MathContext for precision.
    * @return x! (x factorial).
    */
   public static BigDecimal factorial(BigDecimal x, MathContext mc) {
       return BigDecimal.ZERO;
   }
}
